# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/searchByName": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search customer by name",
                "description": "Returns one or list of customers with the same name",
                "operationId": "searchCustomersByName",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/searchByEmail": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search customer by email",
                "description": "Returns customer with same email",
                "operationId": "searchCustomersByEmail",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/searchById": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search customer by ID",
                "description": "Returns a customer with same ID",
                "operationId": "searchCustomersById",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/lastday": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get number of customers registered in the last 24 hours",
                "description": "Returns number of customers registered or added in the last day",
                "operationId": "numberOfCustomersRegistered",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/lastweek": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get number of customers registered in the last 7 days",
                "description": "Returns number of customers registered or added and the average in the last week",
                "operationId": "numberOfCustomersRegistered",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/lastthreemonths": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get number of customers registered in the last 90 days",
                "description": "Returns number of customers registered or added and the average in the last three month",
                "operationId": "numberOfCustomersRegistered",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/lastmonth": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get number of customers registered in the last 30 days",
                "description": "Returns number of customers registered or added and the average in the last month",
                "operationId": "numberOfCustomersRegistered",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/lastyear": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get number of customers registered in the last 365 days",
                "description": "Returns number of customers registered or added and the average in the last year",
                "operationId": "numberOfCustomersRegistered",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ],
    "security": [
        []
    ]
}